translate-toolkit: fix installation for latest oe-core

+++ recipes-support/translate-toolkit/translate-toolkit_1.13.0.bb
@@ -12,12 +12,24 @@ SRC_URI[md5sum] = "f9997a0deb3f1767bb20665a47cd2708"
 SRC_URI[sha256sum] = "dcbbf49058e4196a06e988d9dc1e762321ab0d057c4be035b84e3c11353fc2f8"
 
 do_install_append() {
-	# datatdir remains empty -> delete
-    rmdir ${D}${datadir}
-	# this is a bad hack aligning insalled locations - but have not found a better way
-    cp -r ${D}${STAGING_DIR}/${HOST_SYS}/* ${D}
-	staging_base=`echo ${STAGING_DIR} | cut -d "/" -f2`
-    rm -rf ${D}/$staging_base
+    # move files to correct location
+    # 1. /usr/share/lib -> ${libdir}
+    if [ -d ${D}${datadir}/lib ] ; then
+        mkdir -p ${D}/${libdir}
+        cp -r ${D}${datadir}/lib/* ${D}/${libdir}
+        rm -rf ${D}${datadir}/lib
+    fi
+    
+    # if empty datadir -> delete
+    rmdir ${D}${datadir} || true
+
+    # 2. ${D}${STAGING_DIR}/${HOST_SYS} -> ${D}
+    # this is a bad hack aligning installed locations - but have not found a better way
+    if [ -d ${D}${STAGING_DIR}/${HOST_SYS} ] ; then
+        cp -r ${D}${STAGING_DIR}/${HOST_SYS}/* ${D}
+        staging_base=`echo ${STAGING_DIR} | cut -d "/" -f2`
+        rm -rf ${D}/$staging_base
+    fi
 }
 
 RDEPENDS_${PN} += "bash"
